feat: harden persistent cache runtime policy

+++ crates/model-serving/src/qwen3_5/decoder/cache_layout.rs
@@ -26,6 +26,9 @@ pub fn qwen3_5_decoder_cache_layout(
         0,
         qwen3_5_config.head_dimension() as usize,
     ];
+    // The live Qwen kernels retain full-attention key/value and convolution state as BF16.
+    // Persist the execution dtype exactly: a narrower or wider cache dtype would make restore
+    // either numerically different or incompatible with the model's normal forward path.
     let linear_convolution_dimensions = vec![
         1,
         (qwen3_5_config.linear_convolution_kernel_dimension() as usize).saturating_sub(1),
@@ -43,13 +46,13 @@ pub fn qwen3_5_decoder_cache_layout(
                 DecoderCacheLayerLayout::append_only_attention(
                     DecoderCacheTensorLayout::sequence(
                         QWEN_ATTENTION_KEYS_TENSOR_ROLE,
-                        DecoderCacheTensorDtype::Float32,
+                        DecoderCacheTensorDtype::BFloat16,
                         full_attention_key_value_dimensions.clone(),
                         2,
                     ),
                     DecoderCacheTensorLayout::sequence(
                         QWEN_ATTENTION_VALUES_TENSOR_ROLE,
-                        DecoderCacheTensorDtype::Float32,
+                        DecoderCacheTensorDtype::BFloat16,
                         full_attention_key_value_dimensions.clone(),
                         2,
                     ),
@@ -59,10 +62,12 @@ pub fn qwen3_5_decoder_cache_layout(
                 DecoderCacheLayerLayout::composite(vec![
                     DecoderCacheLayerLayout::recurrent_tensor(DecoderCacheTensorLayout::fixed(
                         QWEN_CONVOLUTION_TENSOR_ROLE,
-                        DecoderCacheTensorDtype::Float16,
+                        DecoderCacheTensorDtype::BFloat16,
                         linear_convolution_dimensions.clone(),
                     )),
                     DecoderCacheLayerLayout::recurrent_tensor(DecoderCacheTensorLayout::fixed(
+                        // The gated-delta recurrent accumulator is Float32 in live execution;
+                        // preserve it independently from the BF16 activation state above.
                         QWEN_RECURRENCE_TENSOR_ROLE,
                         DecoderCacheTensorDtype::Float32,
                         linear_recurrent_dimensions.clone(),
